tests/requirements/dj42_cms41.txt
-r base.txt

Django>=4.2,<5.0
django-cms>=4.1.0,<5.0
djangocms-versioning>=2.0.2
djangocms-text
